+++ models/agol.js
@@ -109,6 +109,8 @@ var AGOL = function (koop) {
   }
 
   agol.updateIfExpired = function (info, options, callback) {
+    // don't try to update things while they are processing
+    if (info.status === 'Processing') return callback(null, info)
     agol.cache.checkExpiration(info, options.layer, function (err, expired) {
       if (err || !expired) return callback(err)
       agol.updateResource(info, options, function (err, status) {