+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+### Fixed
+* Remove possible exception in datasets serialization
+
 ## [3.12.2] - 2016-10-20
 ### Fixed
 * Clean up listeners after import service completes