+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 * Refactor export functions into separate model
 * Always update exports if lastEditDate changes
 * Enqueue exports directly instead of using HTTP after import
+* Wait until correct imported file is ready before enqueueing exports
 
 ### Fixed
 * Remove possible exception in datasets serialization